requirements.txt
alembic==1.5.5
aniso8601==9.0.0
appdirs==1.4.4
attrs==20.3.0
bcrypt==3.2.0
black==20.8b1
cffi==1.14.5
click==7.1.2
flake8==3.8.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.7.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
importlib-metadata==3.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
regex==2020.11.13
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uuid==1.30
Werkzeug==0.16.1
zipp==3.4.0
